XCTAssertEqual doesn't require the args to be the exact same type the way OCUnit did

+++ Example/AlertViewActionSheetExampleTests/ViewControllerTestPlainXCTest.m
@@ -41,7 +41,7 @@
 {
     NSArray *actions = [sut.showAlertButton actionsForTarget:sut forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual([actions count], (NSUInteger)1);
+    XCTAssertEqual([actions count], 1);
     XCTAssertEqualObjects(actions[0], @"showAlert:");
 }
 
@@ -57,13 +57,13 @@
     
     [sut showAlert:nil];
 
-    XCTAssertEqual(alertVerifier.showCount, (NSUInteger)1);
+    XCTAssertEqual(alertVerifier.showCount, 1);
     XCTAssertEqualObjects(alertVerifier.title, @"Get Driving Directions");
     XCTAssertEqualObjects(alertVerifier.message, @"Continue to the Maps app for driving directions?");
     XCTAssertEqual(alertVerifier.delegate, sut);
     NSArray *otherButtonTitles = alertVerifier.otherButtonTitles;
     XCTAssertEqualObjects(alertVerifier.cancelButtonTitle, @"Cancel");
-    XCTAssertEqual([otherButtonTitles count], (NSUInteger)1);
+    XCTAssertEqual([otherButtonTitles count], 1);
     XCTAssertEqualObjects(otherButtonTitles[0], @"OK");
 }
 
@@ -78,7 +78,7 @@
 {
     NSArray *actions = [(sut.showActionSheetButton) actionsForTarget:sut forControlEvent:UIControlEventTouchUpInside];
 
-    XCTAssertEqual([actions count], (NSUInteger)1);
+    XCTAssertEqual([actions count], 1);
     XCTAssertEqualObjects(actions[0], @"showActionSheet:");
 }
 
@@ -94,13 +94,13 @@
     
     [sut showActionSheet:nil];
 
-    XCTAssertEqual(sheetVerifier.showCount, (NSUInteger)1);
+    XCTAssertEqual(sheetVerifier.showCount, 1);
     XCTAssertEqual(sheetVerifier.parentView, [sut view]);
     XCTAssertEqualObjects(sheetVerifier.title, @"http://qualitycoding.org");
     XCTAssertEqual(sheetVerifier.delegate, sut);
     XCTAssertEqualObjects(sheetVerifier.cancelButtonTitle, @"Cancel");
     NSArray *otherButtonTitles = sheetVerifier.otherButtonTitles;
-    XCTAssertEqual([otherButtonTitles count], (NSUInteger)3);
+    XCTAssertEqual([otherButtonTitles count], 3);
     XCTAssertEqualObjects(otherButtonTitles[0], @"Open in Safari");
     XCTAssertEqualObjects(otherButtonTitles[1], @"Copy link");
     XCTAssertEqualObjects(otherButtonTitles[2], @"Separately added button");
